further scaffolding

+++ src/arg_parser.h
@@ -5,7 +5,7 @@ typedef struct Arguments {
   size_t processCount;
   size_t processCountCPUBound;
   size_t seed;
-  int distributionLambda;
+  size_t distributionLambda;
   size_t randomNumberUpperBound;
   size_t contextSwitchMillis;
   double burstTimeAlpha;
